+++ src/app/templates/SortPopup/index.tsx
@@ -1,2 +1,3 @@
 export * from './SortPopup';
 export * from './SortPopup.types';
+export * from './components/SortList';
